added polling function to inbox viewmodel

+++ app/src/main/java/ir/hadiagdamapps/e2eemessenger/data/models/LocalMessageModel.kt
@@ -1,7 +1,7 @@
 package ir.hadiagdamapps.e2eemessenger.data.models
 
 data class LocalMessageModel (
-    val messageId: Int, // don't know if needed
+    val messageId: Long, // don't know if needed
     val conversationId: Int,
     val text: String, // contains both encrypted and decrypted text
     val timestamp: Long,
